some more fixes

+++ Eddy/Components/Setup/CompSimpleFoam.cs
@@ -84,6 +84,17 @@ namespace Eddy
             pManager.AddGenericParameter("Simulation result", "Res", "Eddy simulation result", GH_ParamAccess.item);
         }
 
+
+
+        bool canRun = true;
+        public void taskComplete(object sender, System.EventArgs e)
+        {
+            //RhinoApp.WriteLine("Sim complete");
+            canRun = false;
+            this.ExpireSolution(true);
+        }
+
+
         /// <summary>
         /// This is the method that actually does the work.
         /// </summary>
@@ -271,22 +282,31 @@ namespace Eddy
             DA.GetData("Run Simulation", ref runSimulation);
             DA.GetData("Run Meshing", ref runMeshing);
 
-            if (runMeshing == true)
+
+            if (runMeshing == true && runSimulation == true && canRun)
             {
                 Utilities.DeletePhi(MeshSettings, DOM);
-                Utilities.StartProcess.StartProcessCMDNT("", false, true, false, true, baseWorkingDirectory + @"\run_mesh.bat");
+                Utilities.StartProcess.StartProcessCMDNT("", false, true, false, true, baseWorkingDirectory + @"\run.bat", taskComplete);
             }
 
-            if (runSimulation == true)
+            else if  (runMeshing == true && runSimulation == false && canRun)
             {
                 Utilities.DeletePhi(MeshSettings, DOM);
-                Utilities.StartProcess.StartProcessCMDNT("", false, true, false, true, baseWorkingDirectory + @"\run_sim_all.bat");
+                Utilities.StartProcess.StartProcessCMDNT("", false, true, false, true, baseWorkingDirectory + @"\run_mesh.bat", taskComplete);
+            }
+
+            else if (runMeshing == false && runSimulation == true && canRun)
+            {
+                Utilities.DeletePhi(MeshSettings, DOM);
+                Utilities.StartProcess.StartProcessCMDNT("", false, true, false, true, baseWorkingDirectory + @"\run_sim_all.bat", taskComplete);
             }
 
             #endregion START PROCESSES
 
             OFResult RES = new OFResult(DOM, RunSettings, MeshSettings, baseWorkingDirectory);
             DA.SetData(0, RES);
+
+            canRun = true;
         }
 
         /// <summary>
